+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 1.4.0 (Unreleased)
+
+# Features additions:
+
+ - Added `pools` attribute to `bigip_gtm_wideip` resource for associating GTM pools with a WideIP
 ## 1.3.0 (July 23, 2020)
 
 # Features additions: